space-time bifurcation added
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@3395 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ physics/damagemodels/spacetimefiberbasedfixedcrack.cpp
@@ -183,7 +183,7 @@ bool SpaceTimeFiberBasedFixedCrack::fractured(int direction) const
     {
         return state[0] >= thresholdDamageDensity ;
     }
-    if(direction < 0 || direction >= state.size())
+    if(direction < 0 || direction >= (int) state.size())
     {
         return this->fractured(0) || this->fractured(1) ;
         return getState().max() >= thresholdDamageDensity ;
